Changed the value returned by the method getCloudletFinishedSoFar of Cloudlet class when the index attribute is -1.
The method getCloudletFinishedSoFar was returning the cloudletLength when the index attribute was equal to -1. The documentation of the index indicates that if it is -1, the cloudlet was not executed yet. The method isFinished confirms that, once which the index is equals to -1, it does not compute how many MIPS are remaining for the cloudlet to finish, just returning that it has not finished (because obviously it has not started yet).

Test cases were included.