Changed the return value of the VmAllocationPolicy.optimizeAllocation method from List<Map<String, Object>> that was completely strange and doesn't correctly uses generics. The method return was a List of Maps where the keys were either the strings "vm" or "host". If the key was the string "vm", it meant the value was a Vm instance. If the key was the string "host", it meant the value was a Host instance. Thus, each Map represented the relation between a Vm and the Host where it was to be placed.
This design was very confusing, requiring a lot of explanation and several unsafe typecasts. Now the method (and all related ones that were returning a List<Map<String, Object>>) is returning a Map<Vm, Host>. 

The method DatacenterSimple.processVmMigrate that expected the ev.getData() was a Map<String, Object>, now expects a Map.Entry<Vm, Host>, making to code clearer and reducing typecasts.