THE FIRST CRITICAL CODE REMAKE: UPDATE OF NETWORK RELATED CLASSES -----------------------------------------------------------------
The classes related to the examples at the org.cloudbus.cloudsim.network.datacenter package of the
cloudsim-examples project were almost totally remade.

The network related classes such as AppCloudlet, NetworkCloudlet, NetDatacenterBroker had an extremely amount of duplicated code.
The NetDatacenterBroker was a copy of the original DatacenterBroker class, instead of extending it and overriding
the necessary methods. It was noticed that a lot of code in the created class was exactly equals to the DatacenterBroker,
what shows that such methods should be inherited from the already existing class.

The code that in fact belonged only to the NetDatacenterBroker class was completely repetitive and confuse.
There were very long methods that should be divided into smaller and specific ones in order to avoid
code redundancy and make clear what the code does.

The class used a lot of hard-coded values for creating objects such as Cloudlets and VMs,
that doesn't make sense once these values have to be defined by the developer creating
the simulation. It was clear that the values were defined just to perform the simulations
presented at the paper "NetworkCloudSim: Modelling Parallel Applications in Cloud Simulations",
<http://dx.doi.org/10.1109/UCC.2011.24>. By this way, they aren't useful for other users
that desire to create their own simulations. Further, the creation of hard-coded Cloudlets and VMs inside
the NetDatacenterBroker doesn't make it to be reusable and even violates the **Open/Closed Principle (OCP)**.

Once the mentioned paper doesn't provide low level details about how every class's method works (and it is not supposed that
a paper will present such in deep code details) and that there weren't any unit test for the network related classes,
it was not even feasible to know if the classes and examples were working correctly.
The examples didn't show any useful information, just some vague messages such as "App1, App2, ... AppN" to inform the creation of AppCloudlets,
not providing actual information. In face of that, the best to do was to perform a in-depth refactoring of the classes and examples
in order to provide a actual class hierarchy, eliminate code redundancy and provide short and meaningful methods
that makes clear to understand the code and to just further (unfortunately) introduce test units.

The classes WorkflowApp and BagOfTasksAppCloudlet were deleted because they in fact were just creating hard-coded VMs and Cloudlets
and they weren't providing any specific code. In replacement, they were created the NetworkVmsExampleWorkflowAppCloudlet and NetworkVmsExampleBagOfTasksAppCloudlet classes that just instantiate regular AppCloudlet and NetworkCloudlet classes with specific values in order to provide the desired kind of tasks to be simulated.

The class NetworkConstants was deleted because all of its values were just values used by the examples, and were accordingly moved to the
classes that they belong. Some constants included in this class were just values to be used by Switch related classes. Placing these constants outside the classes that in fact  use them makes the classes design low cohesive. Accordingly, to increase the cohesion, making easy to understand how the elements are related, the constants were moved to the appropriated classes.