Refactoring of Listener related classes and addition of a new set of classes inside the org.cloudbus.cloudsim.listeners package.
A set of interfaces and classes were introduced inside the org.cloudbus.cloudsim.listeners package
to reduce the number of parameters passed to a Listener object when it update method is called
to notify about an expected event.
Now the method just receive an EventInfo object instead of several parameters that
increase the complexity of the Listener.
In some cases, the implementations of this Listener was receiving up to 3 parameters
in the update method. Sometimes when using a Listener, the parameters
of the update method was not clear, mainly when it accepts primitive values
such as double or int. The use of the generic interface EventListener just defines
the type of the generic parameters. The only way to understand what the parameters
are were reading the documentation. Using a specific and single object
as the generic type for the EventListener interface, the usage of the Listener
is completely clear, once the update method receives an object
with specific attributes that can be known at compile time.