Changed License from LGPL to GPL, due to the confusion of CloudSim licensing. The official site (www.cloudbus.org/cloudsim) stated that the version was LGPLv3 for CloudSim 3, but it was changed to Apache v2 in CloudSim 4.
The project at github doesn't has a license file neither for version 3 and 4, as required by every license.
As the source code explicitly includes the http://www.gnu.org/copyleft/gpl.html, that doesn't includes
the version but means that the latest version is covered (as pointed by the link), CloudSim Plus
license was changed to GPLv3.