A functional DatacenterBroker that allows using Java 8 Lambda Expressions to define selection policies (#61)
Refactored the DatacenterBroker interfaces and classes to provide a functional way of defining the policies to select a Datacenter to host a Vm, a fallback Datacenter when the creation of requested VMs failed in the previous selected Datacenter and to select a VM to run a Cloudlet.

The DatacenterBrokerSimple is yet selecting the first Datacenter to place VMs and uses a round-robin policy to select a VM to run the Cloudlets. If such behaviours needs to be changed,
it is not required to create a new DatacenterBroker class. Since there is 3 selection policies to override (the selection of default datacenter, fallback datacenter and VM), the combination of different implementations for these 3 policies will require creation of several DatacenterBroker
implementations that will be impossible to maintain.

As an example, consider that you have 3 different implementations for these 3 policies and you want to make all the possible combinations of these policies. That will require you to
create 27 different implementations of a DatacenterBroker (3 polcies ^ 3 implementations for each one).

Using the new functional implementation there is not need to create a new DatacenterBroker class and the implementations can be exchanged just using the new setDatacenterSupplier,
setFallbackDatacenterSupplier and setVmMapper methods, passing a Lambda Expression to them.