Closes #62 ==========
Changed the UtilizationModel interface adding an enum called Unit that enables the model to define resource usage in percentage or absolute value.
All the already existing UtilizationModels inherited from CloudSim project use a percentage value.

The CloudSim Plus UtilizationModelArithmeticProgression allows changing the unit attribute to any value defined in the Unit enum.
All the code that was getting the resource usage from an UtilizationModel was updated to check what is the unit (if percentage or absolute values),
to correctly compute resource usage.

There may be Cloudlets using UtilizationModels with different units, even inside the same VM.
Depending on where the UtilizationModel is being used, all utilization values are converted to either percentage or absolute values.
The documentation was properly updated to make clear these changes.

Some methods in classes such as CloudletScheduler that had an abstract and confusing name that didn't make clear if
the method was returning an absolute or percentage value were renamed to make them clear.
When the method returns a percentage, its name clearly shows that.