Closes #63 ----------
CloudSim OnClockTickListener updated
====================================

CloudSim OnClockTickListener events are now fired just after all events happening at the same time were processed,
ensuring that the Listener inside the developer simulation will get the most updated simulation state

Refactored the UtilizationModelArithmeticProgression
====================================================

The class now registers the first time its getUtilization() method was called in order to compute the time span from the current simulation time to that time.
That was required because, considering that such an Utilization Model was created during simulation execution (for a Cloudlet that also was created in the same way),
the previous implementation was calculating the time span to get the current utilization
considering that the Utilization Model started to be used at simulation time 0.

That caused an issue. Consider that such a model was created just
at the simulation time 10, the initial usage was 0, the usage increment was 1
and that the first time it was called was at time 20. By this way, the computed
current utilization was 20 instead of 10, since the model started to be used
just at time 10.