Updated Issue #7
Changed the UtilizationModelDynamic to pass a clone instance of itself to the given
lambda function that is accountable to update the utilization progress,
in order to pass a read-only copy of such UtilizationModel to avoid
the lambda expression to change it.
And since the lambda usually call the getUtilization() method internally,
that in turns cal the getUtilization(double) that calls the lambda itself
to increment the current utilization, passing the original UtilizationModel instance caused infinity loop.
The read-only copy of the UtilizationModel passed to the lambda
avoids that issue.