Closes #100
Renamed the DatacenterBroker getVmsCreatedList() method to getVmCreatedList()
and added a getVmExecList() one.

Now the getVmExecList() method is the one which stores the list of all currently running VMs.
These VMs may or may not be running Cloudlets, but they are available
for new submitted Cloudlets to be placed into them.

The old getVmCreatedList() stores all created VMs for the broker
along the entire simulation time.
This way, the list of all created VMs can be queried after the simulation finishes.