Changes confusion in documentation of energy consumption.
Methods `getPower()` and `getPowerInKWattsHour()` informed the
return value was in Watts/Second and Kilowatts/Hour, respectively.
However, the correct units are watt-second (Ws) and Kilowatt-hour (kWh).

Watt is a unit of power, while energy is measured in Ws or kWh.
1 watt-second is the same as 1 Joule/second. However, since
people are more familiar with Ws and kWh, such units were used
instead of Joule/second.

This way, variables named `power` in fact mean `energy`.

A PowerAware interface was introduced to provide common power methods
for PowerModel and Datacenter.

Signed-off-by: Manoel Campos <manoelcampos@gmail.com>