Removes Cloudlet migration code from DatacenterSimple since it doesn't make sense to migration applications and in fact, Cloudlet migration has never been used in fact.
In practice, it's not reasonable to migrate apps autonomously,
since there is all the configurations and dependencies
that would need to be tracked to be migrated too.
That is way we migrate VMs, since they are like black
boxes that we don't have to deal with internal
configurations.

The only kind of apps that is reasonable to migrate
are containers, since they encapsulate configurations
and dependencies as VMs do.

Signed-off-by: Manoel Campos <manoelcampos@gmail.com>