Removes sort from VmAllocationPolicies to improve performance.
The sort was used just to ensure that for the same scenario,
the same Host is always selected for a VM
(for instance, when there are 2 Hosts with the same configuration
that could be selected for a VM, they are sorted according to their IDs).

This sort doesn't have any practical benefit and just drastically
decreases VM allocation performance.

Signed-off-by: Manoel Campos <manoelcampos@gmail.com>