Closes #85 - Fix regression in the issue.
The feature #150 introduced a regression in this issue.
The added Cloudlet.absLength() method was using
returning the absolute value of the length attribute
instead of calling the getLength() method.
Since the latter is overridden in subclasses such
as the NetworkCloudlet, the absLength() method needs
to call the getLength() method instead of using the attribute
directly.

Signed-off-by: Manoel Campos <manoelcampos@gmail.com>