Improves log messages
When a VmAllocationPolicy tries to find a Host for a VM
using the Host.isSuitableForVm method (that calls the same
method on the VmScheduler and ResourceProvisioners),
the VmScheduler method prints an error message
when the managed Host is not suitable for that VM.

That message is misleading. It makes the researcher think that
the VmAllocationPolicy really tried to place the VM into that Hosts,
while it's just inquiring the VmAllocationPolicy to check
if a given Host is suitable. It it's not,
the next Host is checked.
When all hosts are tried and no one is suitable for the VM,
then an error message is already shown.
This way, this intermediate error message was removed.

Signed-off-by: Manoel Campos <manoelcampos@gmail.com>