Fix #202 :: Avoids unnecessary CloudSimTags.VM_UPDATE_CLOUDLET_PROCESSING events
The CloudletSchedulerAbstract.updateProcessing may return 0
to indicate that all running Cloudlets finished but there are
Cloudlets into the waiting list (instead of retuning Double.MAX_VALUE
to indicate there isn't any Cloudlet at all).

In such situation, this value was not being considered and
since zero is smaller than the Simulation's minTimeBetweenEvents,
the value of this attribute was being added to the value
returned by the mentioned method.

The return value indicates the time the next event should be processed,
while, in the mentioned scenario, it was not needed to send
one more VM_UPDATE_CLOUDLET_PROCESSING event.

Furthermore, finished Cloudlets were just being removed from the
execList after updating the processing of Cloudlets in such a List.
The addCloudletsToFinishedList() was called earlier to enable moving
those Cloudlets to the finishedList, then just update the processing
of Cloudlets actually running.

Finally, when the Double.MAX_VALUE was returned
from the VmSimple.updateProcessing,
the decimal places from the current simulation clock were being removed
from the mentioned value (that was required to fix issue #153).
This way, after the method call, when checking
if the returned value was equal to Double.MAX_VALUE,
to indicate there are no more Cloudlets, it wasn't in fact.

Subtracting the simulation's clock decimal places from the returned value
was causing more events to be sent. This was the major performance issue.

Signed-off-by: Manoel Campos <manoelcampos@gmail.com>