Fix #207
When sorting a list of objects (such as Hosts),
if two objects are equal, the Comparator must always
indicate that to avoid
"IllegalArgumentException: Comparison method violates its general contract!".

If the two objects are not equal, then the Comparator has to decide
how to sort them. At that point, the objects can be sorted even at random.

Signed-off-by: Manoel Campos <manoelcampos@gmail.com>