Updates VmAllocationPolicyFirstFit and introduces VmAllocationPolicyRoundRobin.
= VmAllocationPolicyFirstFit

The VmAllocationPolicyFirstFit now defines the max number of tries to find a Host.
The number of iterations of the loop is defined
as the number of Hosts, but as soon a suitable Host is found,
the method returns immediately.

If the lastHostIndex is greater than 0 when the loop starts and
the end of the Host List is reached without finding a suitable Host,
the small change enables the policy to start from the beginning of the
list a keep trying to find a Host, until the number of iterations
reaches the number of Hosts.

= VmAllocationPolicyRoundRobin

This new policy follows the same approach presented above,
defining the max number of tries to find a suitable Host,
however it always moves to the next Hosts after selecting a suitable one.

Signed-off-by: Manoel Campos <manoelcampos@gmail.com>