Closes #218
When a line in a PlanetLab trace file contains 0,
to represent no CPU utilization for a specific time,
such a value was making simulations with indefinite-length Cloudlets
to finish unexpectedly.

If a Cloudlet is not using any CPU, no MIPS are allocated to it.
This way, it wasn't possible to compute the expected finish time,
causing the issue.

Signed-off-by: Manoel Campos <manoelcampos@gmail.com>