Closes #218
When a Cloudlet processing is going to be updated and the
simulation is finishing, if the CPU utilization is zero,
the Cloudlet won't process any instruction at that time.
In such situations, the updateProcessing method was returning
without even sending the CLOUDLET_FINISH message so that
a Cloudlet with negative length could be properly finished.

If the CPU utilization is zero, now the method just returns immediately
if it isn't time to finish the simulation yet.

Signed-off-by: Manoel Campos <manoelcampos@gmail.com>