Fix #223
When all Cloudlets on the execution list were finished and
there were Cloudlets on the waiting List,
the CloudletSchedulerAbstract.updateProcessing method
was retuning Double.MAX_VALUE to indicate that there was no
Cloudlet left to be processed.

In such a case, the method must return the predicted
finish time of Cloudlet from the waiting list
(which was moved to the execution list) that is expected
to finish earlier.

Signed-off-by: Manoel Campos <manoelcampos@gmail.com>