Improved network and power model required for LEAF
1. Improved network model
The existing network topology was highly coupled to BRITE topology files, had an unhandy interface and a O(n^3) worst-case complexity for finding shortest paths, implementing only the Floyd–Warshall algorithm. It was replaced with a network topology based on JGraphT. Users can now make use of the many features and algorithms implemented in this library.

2. Improved power model
The existing power models were only working with data center hosts and were impacting the performance of hosts. They were replaced with a general power model that can be used for hosts and network links but is entirely decoupled from their implementation.