Fix #232
Host.removeMigratingInVm() don't have to deallocate VM resources because
that is already made inside the DatacenterSimple.finishVmMigration().
The Host.removeMigratingInVm() already sets the vm.inMigration to false,
so the DatacenterSimple.finishVmMigration() don't need to do it again.

Re-arrange the code of the DatacenterSimple.finishVmMigration()
so that a VM creation ACK is just sent when the VM is actually
allocated into the target Host.

The VM is re-inserted into the broker vmExecList, because
it's removed when the VM is destroyed into the source Host.

Signed-off-by: Manoel Campos <manoelcampos@gmail.com>