Fix #234
Removes the filter that was selecting just Hosts that have a PowerModel,
where the power consumption is greater than zero.
Hosts can be selected based on power efficient, but they can't be excluded
(filtered) according to that. Even if a Host is power inefficient,
if there is no other suitable one, that Host should be selected
if it has enough physical resources to place a VM.

Signed-off-by: Manoel Campos <manoelcampos@gmail.com>