Fix #235
Creates a clone of the VM which needs to be temporarily placed into a Host
to check if it won't be overload after the placement.
The issue was because calling host.destroyTemporaryVm
was supposed to only destroy the temporary VM created.
But since an actual VM was passed to host.createTemporaryVm,
that actual VM was set as not created after host.destroyTemporaryVm.
Creating a clone of the actual VM so that the clone is temporarily created,
fixed the issue.

This is the kind of issue that happens due to #94.

Signed-off-by: Manoel Campos <manoelcampos@gmail.com>