Prepares for #170
Redesign Host storage to use an actual HarddriveStorage
instead of just a Storage (which is now called SimpleStorage).
This way, we can define disk parameters such as reading speed
that is considered when virtual memory is required
by Cloudlets running inside a VM on the Host.

The FileStorage interface implemented by HarddriveStorage
is now simpler, since we don't want to simulate individual file storage
for hard drive, only for SanStorage.

Adds new Host constructor to accept a HarddriveStorage instead of a long
for defining the storage device.

Since VM and Host have different storage devices now
(SimpleStorage and HarddriveStorage, respectively),
the AbstractMachine interface now is generic,
to indicate the kind of storage to be used by
subclasses.

Signed-off-by: Manoel Campos <manoelcampos@gmail.com>