Fix #368
After implementing virtual memory and BW over-subscription (#170),
if a Cloudlet request more RAM or BW than there is available,
its execution is delayed.

Since the example was making all cloudlets inside a VM to request 100% of RAM and BW,
some of them won't get any RAM or BW to execute, causing wrong execution times.
In such a case, a different UtilizationModel such as the UtilizationModelDynamic
should be created for RAM and BW, indicating that concurrent cloudlets
will use only a fraction of those resources.

Signed-off-by: Manoel Campos <manoelcampos@gmail.com>