Fix #392
When the VmAbstract.setHorizontalScaling, setRamVerticalScaling, setBwVerticalScaling or setPeVerticalScaling
were called, those methods called the validateAndConfigureVmScaling() to configure the VmScaling,
which in turn calls vmScaling.setVm().
This last method tries to get the Resource object from the VM that needs to be scaled.
However, the list of Resource objects is just set when the VM is created.
This way, when the VmScaling tried to get such a resource from the Vm,
it was getting a ResourceNull object instead.

Now, the validateAndConfigureVmScaling() method is just called for each
one of the 4 different types of VmScaling objects when the Vm is created.

Signed-off-by: Manoel Campos <manoelcampos@gmail.com>